requirements.txt
psutil>=5.9.0
pywin32>=306
reportlab>=4.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
PyPDF2>=3.0.0